Rename PluginComponent to KarrotCompnent
