+++ src/base/components/KarrotComponent.vue
