+++ src/group/pages/Wall.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="wrapper">
-    <PluginComponent name="WallHeader" />
+    <KarrotComponent name="WallHeader" />
     <div class="notices">
       <div v-if="joinedActivities.length > 0">
         <JoinedActivities :activities="joinedActivities" />
@@ -25,7 +25,7 @@ import { useCurrentGroupService } from '@/group/services'
 import { useStatusService } from '@/status/services'
 import { newDateRoundedTo5Minutes } from '@/utils/queryHelpers'
 
-import PluginComponent from '@/base/components/PluginComponent.vue'
+import KarrotComponent from '@/base/components/KarrotComponent.vue'
 import AvailableActivities from '@/group/components/AvailableActivities.vue'
 import FeedbackNotice from '@/group/components/FeedbackNotice.vue'
 import JoinedActivities from '@/group/components/JoinedActivities.vue'
@@ -33,7 +33,7 @@ import WallConversation from '@/messages/components/WallConversation.vue'
 
 export default {
   components: {
-    PluginComponent,
+    KarrotComponent,
     JoinedActivities,
     AvailableActivities,
     WallConversation,
