+++ src/sidenav/components/Sidenav.vue
@@ -1,13 +1,13 @@
 <template>
-  <PluginComponent name="Sidenav">
+  <KarrotComponent name="Sidenav">
     <SidenavMap v-if="!$q.platform.is.mobile" />
     <SidenavGroup />
     <SidenavPlaces />
-  </PluginComponent>
+  </KarrotComponent>
 </template>
 
 <script setup>
-import PluginComponent from '@/base/components/PluginComponent.vue'
+import KarrotComponent from '@/base/components/KarrotComponent.vue'
 
 import SidenavGroup from './SidenavGroup.vue'
 import SidenavMap from './SidenavMap.vue'
